recipe입니다...

+++ src/main/java/com/zipchelin/model/dto/recipe/RecipeAddRequestDto.java
@@ -14,7 +14,7 @@ import javax.validation.constraints.NotBlank;
 public class RecipeAddRequestDto {
 
     @NotBlank
-    private int recNum;
+    private long recNum;
 
     @NotBlank
     private int cateNum;
